Move HUB_API_ROOT to hub/utils.py
