+++ ultralytics/hub/auth.py
@@ -1,7 +1,6 @@
 import requests
 
-from ultralytics.hub.config import HUB_API_ROOT
-from ultralytics.hub.utils import request_with_credentials
+from ultralytics.hub.utils import HUB_API_ROOT, request_with_credentials
 from ultralytics.yolo.utils import is_colab
 
 API_KEY_PATH = "https://hub.ultralytics.com/settings?tab=api+keys"
